+++ test/environment_test.rb
@@ -1,5 +1,7 @@
 require 'test_helper'
 
+module Rails; end
+
 class EnvironmentTest < Minitest::Test
 
   def setup
@@ -31,5 +33,14 @@ class EnvironmentTest < Minitest::Test
     assert_equal '127.0.0.1', backend.host
     assert_equal 1234, backend.port
     assert_equal :datadog, backend.implementation
+  end
+
+  def test_uses_env_when_rails_does_not_respond_to_env
+    assert_equal ENV['ENV'], StatsD::Instrument::Environment.environment
+  end
+
+  def test_uses_rails_env_when_rails_is_available
+    Rails.stubs(:env).returns('production')
+    assert_equal 'production', StatsD::Instrument::Environment.environment
   end
 end